+++ crates/wasm_split_cli/src/js.rs
@@ -11,6 +11,7 @@ use crate::{
 type PrefetchMap = HashMap<String, Vec<String>>;
 pub struct LinkModuleWriter<'p> {
     input_module: &'p InputModule<'p>,
+    input_options: &'p crate::Options<'p>,
     program_info: &'p SplitProgramInfo,
     javascript: String,
     prefetch_map: PrefetchMap,
@@ -21,6 +22,7 @@ impl<'p> LinkModuleWriter<'p> {
         Self {
             program_info,
             input_module: emit_state.input(),
+            input_options: emit_state.input_options(),
             javascript: String::new(),
             prefetch_map: HashMap::new(),
         }
@@ -29,11 +31,21 @@ impl<'p> LinkModuleWriter<'p> {
         self.program_info.canary_export_name()
     }
     fn write_main_import(&mut self, mod_path: &str) -> Result<()> {
-        Ok(writeln!(
-            &mut self.javascript,
-            r#"import {{ initSync }} from "{}";"#,
-            mod_path
-        )?)
+        match self.input_options.target {
+            crate::OutputTarget::Web(_) => writeln!(
+                &mut self.javascript,
+                r#"import {{ initSync }} from "{}";
+"#,
+                mod_path
+            )?,
+            crate::OutputTarget::Bundler(_) => writeln!(
+                &mut self.javascript,
+                r#"import * as __wasm from "{}";
+"#,
+                mod_path
+            )?,
+        }
+        Ok(())
     }
     fn write_get_shared_imports(&mut self, main_shares: &str) -> Result<()> {
         let canary_props = if self.input_module.options.debug_assertions {
@@ -41,14 +53,17 @@ impl<'p> LinkModuleWriter<'p> {
         } else {
             String::new()
         };
+        let main_exports = match self.input_options.target {
+            crate::OutputTarget::Web(_) => "initSync(undefined, undefined)",
+            crate::OutputTarget::Bundler(_) => "__wasm",
+        };
         Ok(write!(
             &mut self.javascript,
             r#"let sharedImports = undefined;
 function getSharedImports() {{
     if (sharedImports === undefined) {{
         sharedImports = {{ __wasm_split: {{ {canary_props} }} }};
-        const mainExports = initSync(undefined, undefined);
-        const {{ {main_shares} }} = mainExports;
+        const {{ {main_shares} }} = {main_exports};
         Object.assign(sharedImports.__wasm_split, {{ {main_shares} }});
     }}
     return sharedImports;
@@ -59,12 +74,10 @@ function getSharedImports() {{
     fn write_runtime(&mut self) -> Result<()> {
         self.javascript
             .push_str(include_str!("./snippets/split_wasm.js"));
-        self.javascript
-            .push_str(if self.input_module.options.debug_assertions {
-                include_str!("./snippets/makeFetch.web.debug.js")
-            } else {
-                include_str!("./snippets/makeFetch.web.js")
-            });
+        if self.input_module.options.debug_assertions {
+            self.javascript
+                .push_str(include_str!("./snippets/instantiate.debug.js"));
+        }
         Ok(())
     }
     fn write_export_const(&mut self, name: &str, def: &impl std::fmt::Display) -> Result<()> {
@@ -73,20 +86,36 @@ function getSharedImports() {{
             "export const {name} = {def};"
         )?)
     }
-    fn fetch_opts<'pth>(&self, empty: bool, file_path: impl 'pth + std::fmt::Display) -> String {
+    fn fetcher<'pth>(&self, empty: bool, file_path: impl 'pth + std::fmt::Display) -> String {
         if empty {
-            "undefined".to_string()
+            return "() => async (_imp) => ({})".to_string();
+        }
+        let wrap = if self.input_module.options.debug_assertions {
+            "debugWrap"
         } else {
-            // Note: the expression returned from here should:
-            // - allow lazily fetching the wasm module (no top-level import)
-            // - allow bundlers and downstream code to recognize it as an expression to a path ("relocate" the import)
-            // TODO: try other syntax for different targets:
-            // - `import.source(<file_path>)`
-            // - `URL.resolve(<file_path)` (support is not as good as for new URL)
+            ""
+        };
+        // Note: the expression returned from here should:
+        // - allow lazily fetching the wasm module (no top-level import)
+        // - allow bundlers and downstream code to recognize it as an expression to a path ("relocate" the import)
+        match self.input_options.target {
             // Note: we use the form `new URL(<string literal>, import.meta.url)` which is understood by some
             // bundlers as syntax that can get rewritten if the path from where the file gets fetched is changed
             // (for example due to attaching a has of its contents).
-            format!("new URL({}, import.meta.url)", file_path)
+            crate::OutputTarget::Web(_) => format!(
+                r#"() => {{
+    const src = fetch(new URL({file_path}, import.meta.url));
+    return async (imports) => {wrap}(WebAssembly.instantiateStreaming(src, imports));
+}}
+"#
+            ),
+            crate::OutputTarget::Bundler(_) => format!(
+                r#"() => {{
+    const module = import.source({file_path});
+    return async (imports) => {wrap}(new WebAssembly.Instance(await module, imports));
+}}
+"#
+            ),
         }
     }
     fn write_loaders(&mut self, program: &SplitProgramInfo) -> Result<()> {
@@ -100,10 +129,10 @@ function getSharedImports() {{
             let var_name = format!("__chunk_{module_index}");
             let splits_dbg = splits.iter().cloned().collect::<Vec<_>>().join(", ");
             writeln!(&mut self.javascript, "/* {splits_dbg} */")?;
-            let fetch_opts = self.fetch_opts(is_empty, format_args!("\"./{file_name}.wasm\""));
+            let fetcher = self.fetcher(is_empty, format_args!("\"./{file_name}.wasm\""));
             writeln!(
                 &mut self.javascript,
-                "const {var_name} = makeLoad({fetch_opts}, []);"
+                "const {var_name} = makeLoad({fetcher}, []);"
             )?;
             for split in splits {
                 split_deps
@@ -130,7 +159,7 @@ function getSharedImports() {{
             let loader_name = identifier.loader_name();
             let deps = split_deps.remove(split).unwrap_or_default();
             let deps = deps.join(", ");
-            let fetch_opts = self.fetch_opts(is_empty, format_args!("\"./{file_name}.wasm\""));
+            let fetch_opts = self.fetcher(is_empty, format_args!("\"./{file_name}.wasm\""));
             self.write_export_const(
                 &loader_name,
                 &format_args!("wrapAsyncCb(makeLoad({fetch_opts}, [{deps}]))"),