+++ integration/webpack/js/index.js
@@ -1,0 +1,1 @@
+export { call_args_test } from "../pkg/index.js";