+++ integration/webpack/webpack.config-test.js
@@ -1,0 +1,11 @@
+const config = require('./webpack.config');
+
+module.exports = {
+    ...config,
+    mode: 'development',
+    target: 'node',
+    output: {
+        ...config.output,
+        publicPath: "/",
+    }
+};