+++ test-runner/src/bin/wasm-bindgen-cli-wrapper.rs
@@ -1,0 +1,6 @@
+pub fn main() -> Result<(), exec::Error> {
+    Err(exec::execvp(
+        env!("CARGO_BIN_FILE_WASM_BINDGEN_CLI_wasm-bindgen"),
+        std::env::args(),
+    ))
+}